Updates re Two Dims in one
New Per_Tot hierarchy with consolidations for the same period across all years allows a view in PAX showing Periods on Columns and Years on Rows. Also make the separator in Financial Years at YYYY-YY customisable as the default of YYYY/YY does not play well when used in a file name.